Split pathutils into two utility files
