+++ src/components/manager.ts
@@ -6,6 +6,8 @@ import * as chokidar from 'chokidar'
 import * as micromatch from 'micromatch'
 import type {latexParser} from 'latex-utensils'
 import * as utils from '../utils/utils'
+import {PathRegExp} from '../utils/inputfilepath'
+import type {MatchPath} from '../utils/inputfilepath'
 
 import type {Extension} from '../main'
 import * as eventbus from './eventbus'
@@ -17,8 +19,7 @@ import type {ILwCompletionItem} from '../providers/completer/interface'
 import {PdfWatcher} from './managerlib/pdfwatcher'
 import {BibWatcher} from './managerlib/bibwatcher'
 import {FinderUtils} from './managerlib/finderutils'
-import {PathUtils, PathRegExp} from './managerlib/pathutils'
-import type {MatchPath} from './managerlib/pathutils'
+import {PathUtils} from './managerlib/pathutils'
 import {IntellisenseWatcher} from './managerlib/intellisensewatcher'
 
 import {Mutex} from '../lib/await-semaphore'
