+++ src/providers/structure.ts
@@ -1,10 +1,11 @@
 import * as vscode from 'vscode'
 import * as path from 'path'
-import {latexParser,bibtexParser} from 'latex-utensils'
+import { latexParser,bibtexParser } from 'latex-utensils'
 
 import type { Extension } from '../main'
-import * as utils from '../utils/utils'
-import { PathRegExp, MatchPath, MatchType } from '../components/managerlib/pathutils'
+import { resolveFile } from '../utils/utils'
+import { PathRegExp, MatchType } from '../utils/inputfilepath'
+import type { MatchPath } from '../utils/inputfilepath'
 
 export class SectionNodeProvider implements vscode.TreeDataProvider<Section> {
 
@@ -302,7 +303,7 @@ export class SectionNodeProvider implements vscode.TreeDataProvider<Section> {
         if (['input', 'InputIfFileExists', 'include', 'SweaveInput',
              'subfile', 'loadglsentries'].includes(node.name.replace(/\*$/, ''))
             && cmdArgs.length > 0) {
-            candidate = utils.resolveFile(
+            candidate = resolveFile(
                 [path.dirname(file),
                  path.dirname(this.extension.manager.rootFile || ''),
                  ...texDirs],
@@ -311,7 +312,7 @@ export class SectionNodeProvider implements vscode.TreeDataProvider<Section> {
         // \import{sections/}{section1.tex}
         if (['import', 'inputfrom', 'includefrom'].includes(node.name.replace(/\*$/, ''))
             && cmdArgs.length > 1) {
-            candidate = utils.resolveFile(
+            candidate = resolveFile(
                 [cmdArgs[0],
                  path.join(
                     path.dirname(this.extension.manager.rootFile || ''),
@@ -321,7 +322,7 @@ export class SectionNodeProvider implements vscode.TreeDataProvider<Section> {
         // \subimport{01-IntroDir/}{01-Intro.tex}
         if (['subimport', 'subinputfrom', 'subincludefrom'].includes(node.name.replace(/\*$/, ''))
             && cmdArgs.length > 1) {
-            candidate = utils.resolveFile(
+            candidate = resolveFile(
                 [path.dirname(file)],
                 path.join(cmdArgs[0], cmdArgs[1]))
         }
